added Validator class for validating responses

+++ src/main/java/com/serg/bash/monitor/dto/Url.java
@@ -15,7 +15,7 @@ public class Url {
 
     private int periodMonitoring;
 
-    private int responseTime;
+    private long responseTime;
 
     private int responseCode;
